test(chaos): address compact-claim backdate review findings
The kind-filtered backdate aged the shared batch claimed_at while only
targeted members received stale attempt_state rows; an untargeted open
member with no attempt_state row yet inherited staleness through
COALESCE(attempt_state.heartbeat_at, claimed_at). Untargeted open
members now get their effective heartbeat pinned to now() (DO NOTHING
preserves fresher rows) before the cursor moves.

The open-member predicate also gains the canonical done_entries /
deferred_jobs / dlq_entries exclusions: completion closes compact
members through those tables, not lease_claim_closures, so terminal
members could previously inflate the affected-member count or age
batch cursors. The cursor UPDATE now requires a matching OPEN member
too. Also clamp the paced-cooldown computation in u128 before
narrowing.

+++ awa-worker/src/maintenance.rs
@@ -327,8 +327,8 @@ impl MaintenanceBranchTracker {
                 return 0;
             }
             let interval_ms = tick_interval.as_millis().max(1);
-            let ticks = (2 * last_duration.as_millis()).div_ceil(interval_ms) as u32;
-            ticks.clamp(1, cooldown_ticks)
+            let ticks = (2 * last_duration.as_millis()).div_ceil(interval_ms);
+            ticks.clamp(1, cooldown_ticks as u128) as u32
         };
         if let Some(last_duration) = state.last_duration.take() {
             // Integer-ratio thresholds — avoid f64 in the hot path.
